Add:update example_app page for split to components and functions
