+++ src/pages/01_example_app.py
@@ -1,7 +1,8 @@
-import altair as alt
-import numpy as np
-import pandas as pd
+# 01_example_app.py
 import streamlit as st
+
+from functions.calculations import calculate_spiral
+from components.spiral_chart import spiral_chart
 
 """
 # Welcome to Streamlit!
@@ -18,29 +19,8 @@ below is an example of what you can do with just a few lines of code:
 num_points = st.slider("Number of points in spiral", 1, 10000, 1100)
 num_turns = st.slider("Number of turns in spiral", 1, 300, 31)
 
-indices = np.linspace(0, 1, num_points)
-theta = 2 * np.pi * num_turns * indices
-radius = indices
+# calculate axises and indices
+x, y, indices = calculate_spiral(num_points, num_turns)
 
-x = radius * np.cos(theta)
-y = radius * np.sin(theta)
-
-df = pd.DataFrame(
-    {
-        "x": x,
-        "y": y,
-        "idx": indices,
-        "rand": np.random.randn(num_points),
-    }
-)
-
-st.altair_chart(
-    alt.Chart(df, height=700, width=700)
-    .mark_point(filled=True)
-    .encode(
-        x=alt.X("x", axis=None),
-        y=alt.Y("y", axis=None),
-        color=alt.Color("idx", legend=None, scale=alt.Scale()),
-        size=alt.Size("rand", legend=None, scale=alt.Scale(range=[1, 150])),
-    )
-)
+# plot chart
+spiral_chart(x, y, indices, num_points)
